README.md
1. Reliable transfer over reliable channel
This project consists of building an SFTP (Simple File Transfer Protocol). It consists of a client and a server that exchange one file using TCP. The client accepts 4 arguments: the name of the two files (<input> and <output>), and the IP address (or Internet name) and the port number of the server. The server starts first and waits for a connection request. The client requests a connection and then sends the name of the file <output> for output and the data in file <input> to the server, which saves the info received in the file <output>. The client reads the file and sends the data in chunks of 10 bytes. After sending the file, the client closes the connection and exits. The server receives the data and writes the file in chunks of 5 bytes.The server needs to know when the transmission is over so that it can close the file.

2. Reliable transfer over unreliablechannel with bit errors
This project consists of building a Stop and Wait (S&W) reliable protocol. The S&W is going to be built on top of UDP, and it is supposed to provide a reliable transport service to the SFTP application (developed in part 1, which needs to change to call your new send and receive functions). Messages are sent one at a time, and each message needs to be acknowledged when received, before a new message can be sent. The S&W consists of a client and a server. Communication is unidirectional, i.e., data flows from the client to the server. The server starts first and waits for messages. The client starts the communication. Messages have seq number 0 or 1. Before sending each message, a checksum is calculated and added to the S&W header. After sending each message, the client waits for a corresponding ACK. When it arrives, if it is not the corresponding ACK (or if the checksum does not match), the message is sent again. If it is the corresponding ACK, the client changes state and returns to the application, which can now send one more message. This means that the S&W blocks on writes. The server, after receiving a message, checks its checksum. If the message is correct and has the right seq number, the server sends an ACK0 or ACK1 message (according to the seq number) to the client, changes state accordingly, and deliver data to the application. The protocol should deal properly with duplicate data messages and duplicate ACK messages. The S&W message contains the header and the application data. No reordering is necessary, since the S&W is sending the exact message given by the application, one by one.